common : move arg parser to arg.cpp
